.claude/skills/crawl/pyproject.toml
[project]
name = "crawl"
version = "0.1.0"
description = "Web crawling skill"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0"
]